cleanup: isolate query read set construction
